refactored tsx files into appropriate folders
