+++ components/ReposTableComponents/DependenciesContainer.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import CollapsibleTable from "./CollapsibleTable";
-import styles from "../styles/DependenciesContainer.module.css";
-import sharedStyles from "../styles/TreeView.module.css";
+import styles from "../../styles/DependenciesContainer.module.css";
+import sharedStyles from "../../styles/TreeView.module.css";
 import SearchBar from "./SearchBar";
-import { DependencyData } from "../src/dataProcessing";
-import config from "../config.json";
+import { DependencyData } from "../../src/dataProcessing";
+import config from "../../config.json";
 // import { Grid } from "@mui/material"
 //import filterIcon from "../components/images/filter.svg" ;
 import {
@@ -14,7 +14,7 @@ import {
   lastRequest,
   PageLoaderSetData,
   PageLoaderSetLoading,
-} from "./PageLoader";
+} from "../PageLoader";
 
 let refreshing = false;
