+++ components/ReposTableComponents/InverseSubRow.tsx
@@ -1,12 +1,12 @@
 import TableCell from "@mui/material/TableCell";
 import TableRow from "@mui/material/TableRow";
-import { semVerToString } from "../src/semVer";
+import { semVerToString } from "../../src/semVer";
 import { SubRowProps } from "./SubRow";
 import Image from "next/image";
-import styles from "../styles/SubRow.module.css";
-import RedIcon from "./images/redIcon.svg"
-import YellowIcon from "./images/yellowIcon.svg";
-import GreenIcon from "./images/greenIcon.svg";
+import styles from "../../styles/SubRow.module.css";
+import RedIcon from "../images/redIcon.svg"
+import YellowIcon from "../images/yellowIcon.svg";
+import GreenIcon from "../images/greenIcon.svg";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 
 type InverseSubRowProps = {
