+++ components/ReposTableComponents/Row.tsx
@@ -5,14 +5,14 @@ import IconButton from "@mui/material/IconButton";
 import { Tooltip, TableRow, TableHead, TableCell, Table } from "@mui/material";
 import KeyboardArrowRightIcon from '@mui/icons-material/KeyboardArrowRight';
 import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
-import RedIcon from "./images/redIcon.svg";
-import YellowIcon from "./images/yellowIcon.svg";
-import greenIcon from "./images/greenIcon.svg";
+import RedIcon from "../images/redIcon.svg";
+import YellowIcon from "../images/yellowIcon.svg";
+import greenIcon from "../images/greenIcon.svg";
 import Image from "next/image";
 import Tabs from "./Tabs";
-import { semVerToString } from "../src/semVer";
-import styles from "../styles/Row.module.css";
-import { redDef, yellowDef, greenDef } from "./LightStatus";
+import { semVerToString } from "../../src/semVer";
+import styles from "../../styles/Row.module.css";
+import { redDef, yellowDef, greenDef } from "../HelpComponents/LightStatus";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 import QuestionMark from "@mui/icons-material/QuestionMark";
