+++ components/ReposTableComponents/SearchBar.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Image from "next/image";
-import styles from "../styles/SearchBar.module.css";
-import magnifyingGlass from "./images/magnifying-glass.svg";
+import styles from "../../styles/SearchBar.module.css";
+import magnifyingGlass from "../images/magnifying-glass.svg";
 
 export default function SearchBar(props: {
   searchTerm: any;
