+++ components/ReposTableComponents/SubRow.tsx
@@ -1,15 +1,15 @@
 import TableCell from "@mui/material/TableCell";
 import TableRow from "@mui/material/TableRow";
 import Tooltip from "@mui/material/Tooltip";
-import { semVerToString } from "../src/semVer";
-import { PackageData } from "../hooks/useProcessDependencyData";
-import styles from "../styles/SubRow.module.css";
+import { semVerToString } from "../../src/semVer";
+import { PackageData } from "../../hooks/useProcessDependencyData";
+import styles from "../../styles/SubRow.module.css";
 import Image from "next/image";
-import RedIcon from "./images/redIcon.svg";
-import YellowIcon from "./images/yellowIcon.svg";
-import GreenIcon from "./images/greenIcon.svg";
+import RedIcon from "../images/redIcon.svg";
+import YellowIcon from "../images/yellowIcon.svg";
+import GreenIcon from "../images/greenIcon.svg";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
-import { redDef, yellowDef, greenDef } from "./LightStatus";
+import { redDef, yellowDef, greenDef } from "../HelpComponents/LightStatus";
 
 export type SubRowProps = {
   dependency: PackageData;
