+++ components/SummaryComponents/HeaderContainer.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import headerStyles from "../styles/HeaderContainer.module.css";
+import headerStyles from "../../styles/HeaderContainer.module.css";
 // import sharedStyles from "../styles/TreeView.module.css";
-import config from "../config.json";
+import config from "../../config.json";
 import ForestIcon from "@mui/icons-material/Forest";
-import { RepoOverviewCondensedProps } from "./SummaryComponents/CondensedSummary/CondensedSummary";
+import { RepoOverviewCondensedProps } from "./CondensedSummary/CondensedSummary";
 export default function HeaderContainer() {
   return (
     <>
